feat(commandUtils): implement new experimental`spawnAsPromised()` version
The new version relies on the fact that `child_process.spawn()` will be
called with `shell: true`, thus the command does not need special
handling (e.g. to support piping).

The new version is more powerful and may become the default in the
future, but it is still in experimental stage (and can be turned on on a
per-command basis with the `--gkcu-sapVersion=2" flag).

Here be dragons!